label:enhancement adding DoSlowCmd in openserver.py
Hi gents,

I was following this project. I did some personal coding that I believe is good to share with you.
Idea is to add DoSlowCmd for long calculation to avoid timeout of a script.

I followed the orignal VBA code. I did a try and it worked fine. We could go for furhter testing. Can we discuss about this patch?
Tell me your thoughts.

Marc